move some funcs to headers and mark inline
