+++ CMakeLists.txt
@@ -131,13 +131,11 @@ set( ATHELAS_SOURCES
   "src/eos/eos_marshak.cpp"
   "src/eos/eos_paczynski.cpp"
   "src/eos/eos_polytropic.cpp"
-  "src/fluid/fluid_utilities.cpp"
   "src/fluid/hydro_package.cpp"
   "src/heating/nickel_package.cpp"
   "src/history/history.cpp"
   "src/interface/params.cpp"
   "src/radiation/radhydro_package.cpp"
-  "src/radiation/rad_utilities.cpp"
   "src/geometry/grid.cpp"
   "src/gravity/gravity_package.cpp"
   "src/io/io.cpp"
